requirements.txt
Django == 1.9.6
django-phonenumber-field == 1.1.0
djangorestframework == 3.3.3
Pillow == 3.2.0
psycopg2 == 2.6.1
pytest-django
